#3-0 [Add] 약관 데이터 관리 로직 실행 중 관리자 권한 확인

+++ backend/src/main/java/dragonb/bearfamily/backend/service/TermsService.java
@@ -3,6 +3,8 @@ package dragonb.bearfamily.backend.service;
 import java.util.List;
 import java.util.Optional;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,9 +18,10 @@ public class TermsService {
     @Autowired
     TermsRepository termsRepository;
 
-    public Terms getTerms(Long id) throws Exception{
+    @Autowired
+    CommonService commonService;
 
-        
+    public Terms getTerms(Long id) throws Exception{
         Optional<Terms> resultTerms = termsRepository.findById(id);
         if(!resultTerms.isPresent()){
             throw new Exception();
@@ -32,7 +35,11 @@ public class TermsService {
         return termsRepository.findAll();
     }
 
-    public Terms postTerms(TermsDTO termsDTO) throws Exception{
+    public Terms postTerms(TermsDTO termsDTO, HttpServletRequest request) throws Exception{
+        if(!commonService.isAdmin(request)){
+            throw new Exception();
+        }
+
         return termsRepository.save(Terms.builder()
         .title(termsDTO.getTitle())
         .content(termsDTO.getContent())
@@ -40,7 +47,11 @@ public class TermsService {
         .build());
     }
 
-    public Terms putTerms(TermsDTO termsDTO, Long id) throws Exception{
+    public Terms putTerms(TermsDTO termsDTO, Long id, HttpServletRequest request) throws Exception{
+        if(!commonService.isAdmin(request)){
+            throw new Exception();
+        }
+
         Optional<Terms> resultTerms = termsRepository.findById(id);
         if(!resultTerms.isPresent()){
             throw new Exception();
@@ -55,7 +66,11 @@ public class TermsService {
         return test;
     }
 
-    public void deleteTerms(Long id) throws Exception{
+    public void deleteTerms(Long id, HttpServletRequest request) throws Exception{
+        if(!commonService.isAdmin(request)){
+            throw new Exception();
+        }
+
         termsRepository.deleteById(id);
     }
 }
